+++ src/xSTIR/cSTIR/include/sirf/STIR/stir_data_containers.h
@@ -513,6 +513,10 @@ namespace sirf {
 		void semibinary_op(const DataContainer& a_x, float y, float(*f)(float, float));
 		void binary_op(const DataContainer& a_x, const DataContainer& a_y, float(*f)(float, float));
 
+		virtual size_t address() const {
+			THROW("data address defined only for data in memory");
+		}
+
 	protected:
 		static std::string _storage_scheme;
 		static std::shared_ptr<STIRAcquisitionData> _template;
@@ -889,6 +893,12 @@ namespace sirf {
 		virtual bool supports_array_view() const
 		{
 			return STIR_VERSION >= 060200;
+		}
+		virtual size_t address() const {
+			auto *pd_ptr = dynamic_cast<const stir::ProjDataInMemory*>(data().get());
+			if (is_null_ptr(pd_ptr))
+				THROW("address() defined only for data in memory");
+			return reinterpret_cast<size_t>(pd_ptr->get_const_data_ptr());
 		}
 
 	private:
@@ -1393,6 +1403,10 @@ namespace sirf {
 		void unary_op(const DataContainer& a_x, float(*f)(float));
 		void semibinary_op(const DataContainer& a_x, float y, float(*f)(float, float));
 		void binary_op(const DataContainer& a_x, const DataContainer& a_y, float(*f)(float, float));
+
+		size_t address() const {
+		    return reinterpret_cast<size_t>(_data->get_const_full_data_ptr());
+		}
 
 	private:
 		/// Clone helper function. Don't use.